+++ BE/src/main/java/com/codesquad/issuetracker/issue/data/Issue.java
@@ -1,6 +1,8 @@
 package com.codesquad.issuetracker.issue.data;
 
 import com.codesquad.issuetracker.auth.data.User;
+import com.codesquad.issuetracker.issue.data.relation.IssueLabelRelation;
+import com.codesquad.issuetracker.issue.data.relation.IssueMilestoneRelation;
 import com.codesquad.issuetracker.issue.web.model.IssueQuery;
 import com.codesquad.issuetracker.label.data.Label;
 import java.time.LocalDateTime;
@@ -41,11 +43,11 @@ public class Issue {
   private String description;
 
   @Builder.Default
-  @OneToMany(mappedBy = "issue", fetch = FetchType.EAGER, cascade = CascadeType.PERSIST)
+  @OneToMany(mappedBy = "issue", fetch = FetchType.EAGER, cascade = CascadeType.ALL)
   private Set<IssueLabelRelation> issueLabelRelations = new LinkedHashSet<>();
 
   @Builder.Default
-  @OneToMany(mappedBy = "issue", fetch = FetchType.EAGER, cascade = CascadeType.PERSIST)
+  @OneToMany(mappedBy = "issue", fetch = FetchType.EAGER, cascade = CascadeType.ALL)
   private Set<IssueMilestoneRelation> issueMilestoneRelations = new HashSet<>();
 
   @CreationTimestamp
@@ -69,13 +71,28 @@ public class Issue {
         .description(query.getDescription())
         .build();
 
-    issue.issueLabelRelations = labels.stream().map(label -> IssueLabelRelation.of(issue, label))
+    LinkedHashSet<Long> idOfLabels = labels.stream().map(Label::getId)
+        .collect(Collectors.toCollection(LinkedHashSet::new));
+    issue.issueLabelRelations = idOfLabels.stream()
+        .map(id -> IssueLabelRelation.of(issue, id))
         .collect(Collectors.toSet());
 
     return issue;
   }
 
   public Boolean isSameUser(User user) {
-    return this.userId.equals(user.getUserId());
+    return userId.equals(user.getUserId());
+  }
+
+  public Set<Long> idOfLabels() {
+    return issueLabelRelations.stream()
+        .map(IssueLabelRelation::getLabelId)
+        .collect(Collectors.toSet());
+  }
+
+  public Set<Long> idOfMilestones() {
+    return issueMilestoneRelations.stream()
+        .map(IssueMilestoneRelation::getMilestoneId)
+        .collect(Collectors.toSet());
   }
 }
