+++ BE/src/main/java/com/codesquad/issuetracker/label/data/Label.java
@@ -1,19 +1,10 @@
 package com.codesquad.issuetracker.label.data;
 
-import com.codesquad.issuetracker.issue.data.IssueLabelRelation;
 import com.codesquad.issuetracker.label.web.model.LabelQuery;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -34,11 +25,6 @@ public class Label {
   private String title;
   private String description;
   private String color;
-
-  @Builder.Default
-  @JsonInclude(Include.NON_NULL)
-  @OneToMany(mappedBy = "label", fetch = FetchType.EAGER, cascade = CascadeType.PERSIST)
-  private List<IssueLabelRelation> issueLabelRelations = new ArrayList<>();
 
   public static Label of(Long id) {
     return Label.builder()
